refactor: P3-T1 - 拆分 PracticeSessionViewModel 文件结构
