+++ LonelyPianistAVP/ViewModels/PracticeSessionViewModel.swift
@@ -6,7 +6,7 @@ import simd
 @MainActor
 @Observable
 final class PracticeSessionViewModel {
-    private let decisionLogger = Logger(
+    let decisionLogger = Logger(
         subsystem: Bundle.main.bundleIdentifier ?? "LonelyPianistAVP",
         category: "Step3AudioDecision"
     )
@@ -39,19 +39,19 @@ final class PracticeSessionViewModel {
     var autoplayState: AutoplayState = .off
     private(set) var calibration: PianoCalibration?
     private(set) var keyboardGeometry: PianoKeyboardGeometry?
-    private(set) var pressedNotes: Set<Int> = []
+    var pressedNotes: Set<Int> = []
     var feedbackState: VisualFeedbackState = .none
     var isSustainPedalDown = false
-    private(set) var audioRecognitionErrorMessage: String?
+    var audioRecognitionErrorMessage: String?
     private(set) var audioPlaybackErrorMessage: String?
     var autoplayErrorMessage: String?
     var audioErrorMessage: String? {
         audioRecognitionErrorMessage ?? audioPlaybackErrorMessage
     }
 
-    private(set) var audioRecognitionStatus: PracticeAudioRecognitionStatus = .idle
-    private(set) var audioRecognitionDebugSnapshot: PracticeAudioRecognitionDebugSnapshot = .empty
-    private(set) var handGateState = HandGateState(
+    var audioRecognitionStatus: PracticeAudioRecognitionStatus = .idle
+    var audioRecognitionDebugSnapshot: PracticeAudioRecognitionDebugSnapshot = .empty
+    var handGateState = HandGateState(
         isNearKeyboard: false,
         hasDownwardMotion: false,
         exactPressedNotes: [],
@@ -59,20 +59,20 @@ final class PracticeSessionViewModel {
     )
     var noteMatchTolerance: Int = 1
 
-    private let pressDetectionService: PressDetectionServiceProtocol
+    let pressDetectionService: PressDetectionServiceProtocol
     let chordAttemptAccumulator: ChordAttemptAccumulatorProtocol
     let sleeper: SleeperProtocol
     let sequencerPlaybackService: PracticeSequencerPlaybackServiceProtocol
-    private let audioRecognitionService: PracticeAudioRecognitionServiceProtocol?
-    private let audioStepAttemptAccumulator: AudioStepAttemptAccumulator
-    private let handPianoActivityGate: HandPianoActivityGate
+    let audioRecognitionService: PracticeAudioRecognitionServiceProtocol?
+    let audioStepAttemptAccumulator: AudioStepAttemptAccumulator
+    let handPianoActivityGate: HandPianoActivityGate
     private let manualAdvanceModeProvider: () -> ManualAdvanceMode
     var feedbackResetTask: Task<Void, Never>?
     var autoplayTask: Task<Void, Never>?
     var autoplayTaskGeneration = 0
-    private var audioRecognitionEventsTask: Task<Void, Never>?
-    private var audioRecognitionStatusTask: Task<Void, Never>?
-    private var audioRecognitionDebugTask: Task<Void, Never>?
+    var audioRecognitionEventsTask: Task<Void, Never>?
+    var audioRecognitionStatusTask: Task<Void, Never>?
+    var audioRecognitionDebugTask: Task<Void, Never>?
     private(set) var tempoMap = MusicXMLTempoMap(tempoEvents: [])
     private var measureSpans: [MusicXMLMeasureSpan] = []
     var manualReplayTask: Task<Void, Never>?
@@ -89,11 +89,11 @@ final class PracticeSessionViewModel {
     var manualHighlightTransitionTask: Task<Void, Never>?
     var audioRecognitionGeneration = 0
     var isAudioRecognitionRunning = false
-    private var audioRecognitionSuppressUntil: Date?
-    private let audioRecognitionSuppressDuration: TimeInterval = 0.6
-    private let audioRecognitionEnabledSnapshot = MusicXMLRealisticPlaybackDefaults.audioRecognitionEnabled
-    private var practiceAudioRecognitionDetectorModeSnapshot: PracticeAudioRecognitionDetectorMode = .harmonicTemplate
-    private var harmonicTemplateTuningProfileSnapshot: HarmonicTemplateTuningProfile = .lowLatencyDefault
+    var audioRecognitionSuppressUntil: Date?
+    let audioRecognitionSuppressDuration: TimeInterval = 0.6
+    let audioRecognitionEnabledSnapshot = MusicXMLRealisticPlaybackDefaults.audioRecognitionEnabled
+    var practiceAudioRecognitionDetectorModeSnapshot: PracticeAudioRecognitionDetectorMode = .harmonicTemplate
+    var harmonicTemplateTuningProfileSnapshot: HarmonicTemplateTuningProfile = .lowLatencyDefault
 
     init(
         pressDetectionService: PressDetectionServiceProtocol,
@@ -419,52 +419,6 @@ final class PracticeSessionViewModel {
         refreshAudioRecognitionForCurrentState()
     }
 
-    func handleFingerTipPositions(_ fingerTips: [String: SIMD3<Float>], at timestamp: Date = .now) -> Set<Int> {
-        guard let keyboardGeometry else { return [] }
-        let detected = pressDetectionService.detectPressedNotes(
-            fingerTips: fingerTips,
-            keyboardGeometry: keyboardGeometry,
-            at: timestamp
-        )
-        if detected.isEmpty == false {
-            pressedNotes = detected
-            handGateState = handPianoActivityGate.evaluate(
-                fingerTips: fingerTips,
-                keyboardGeometry: keyboardGeometry,
-                exactPressedNotes: detected
-            )
-            if autoplayState == .off, isManualReplayPlaying == false, let currentStep {
-                let expected = uniqueMIDINotes(in: currentStep)
-                let isMatched = chordAttemptAccumulator.register(
-                    pressedNotes: detected,
-                    expectedNotes: expected,
-                    tolerance: noteMatchTolerance,
-                    at: timestamp
-                )
-                if isMatched {
-                    setFeedback(.correct)
-                    if autoplayState == .off {
-                        advanceToNextStep()
-                    }
-                } else {
-                    let unrelatedPressDetected = detected.contains { pressed in
-                        expected.contains(where: { abs($0 - pressed) <= noteMatchTolerance }) == false
-                    }
-                    if unrelatedPressDetected {
-                        setFeedback(.wrong)
-                    }
-                }
-            }
-        } else {
-            handGateState = handPianoActivityGate.evaluate(
-                fingerTips: fingerTips,
-                keyboardGeometry: keyboardGeometry,
-                exactPressedNotes: []
-            )
-        }
-        return detected
-    }
-
     private func advanceToNextManualUnit() {
         guard steps.isEmpty == false else {
             state = .idle
@@ -477,7 +431,7 @@ final class PracticeSessionViewModel {
         moveToStep(nextIndex, shouldPlaySound: autoplayState == .off)
     }
 
-    private func advanceToNextStep() {
+    func advanceToNextStep() {
         guard steps.isEmpty == false else {
             state = .idle
             return
@@ -534,7 +488,7 @@ final class PracticeSessionViewModel {
         stopAudioRecognition()
     }
 
-    private func uniqueMIDINotes(in step: PracticeStep) -> [Int] {
+    func uniqueMIDINotes(in step: PracticeStep) -> [Int] {
         Set(step.notes.map(\.midiNote)).sorted()
     }
 
@@ -543,12 +497,7 @@ final class PracticeSessionViewModel {
         audioPlaybackErrorMessage = audioErrorText(for: error)
     }
 
-    private func recordAudioRecognitionError(_ error: Error) {
-        guard audioRecognitionErrorMessage == nil else { return }
-        audioRecognitionErrorMessage = audioErrorText(for: error)
-    }
-
-    private func audioErrorText(for error: Error) -> String {
+    func audioErrorText(for error: Error) -> String {
         if let localized = error as? LocalizedError, let description = localized.errorDescription,
            description.isEmpty == false
         {
@@ -557,7 +506,7 @@ final class PracticeSessionViewModel {
         return String(describing: error)
     }
 
-    private func setFeedback(_ state: VisualFeedbackState, duration: TimeInterval = 0.25) {
+    func setFeedback(_ state: VisualFeedbackState, duration: TimeInterval = 0.25) {
         feedbackState = state
         feedbackResetTask?.cancel()
         feedbackResetTask = Task { @MainActor [weak self] in
@@ -566,230 +515,5 @@ final class PracticeSessionViewModel {
             guard Task.isCancelled == false else { return }
             feedbackState = .none
         }
-    }
-
-    private func bindAudioRecognitionStreamsIfNeeded() {
-        guard let audioRecognitionService else { return }
-        guard audioRecognitionEventsTask == nil else { return }
-
-        audioRecognitionEventsTask = Task { [weak self] in
-            for await event in audioRecognitionService.events {
-                await MainActor.run {
-                    self?.handleAudioRecognitionEvent(event)
-                }
-            }
-        }
-
-        audioRecognitionStatusTask = Task { [weak self] in
-            for await status in audioRecognitionService.statusUpdates {
-                await MainActor.run {
-                    self?.audioRecognitionStatus = status
-                    if case .permissionDenied = status {
-                        self?.audioRecognitionErrorMessage = "未授予麦克风权限"
-                    }
-                    if case let .engineFailed(reason) = status {
-                        self?.audioRecognitionErrorMessage = reason
-                    }
-                }
-            }
-        }
-
-        audioRecognitionDebugTask = Task { [weak self] in
-            for await snapshot in audioRecognitionService.debugSnapshots {
-                await MainActor.run {
-                    self?.audioRecognitionDebugSnapshot = snapshot
-                }
-            }
-        }
-    }
-
-    func refreshAudioRecognitionForCurrentState() {
-        guard let audioRecognitionService else { return }
-        guard isPracticeAudioRecognitionEnabled else {
-            stopAudioRecognition()
-            return
-        }
-        guard autoplayState == .off else {
-            stopAudioRecognition()
-            return
-        }
-        guard isManualReplayPlaying == false else {
-            stopAudioRecognition()
-            return
-        }
-        guard case .guiding = state, let currentStep else {
-            stopAudioRecognition()
-            return
-        }
-
-        let expectedMIDINotes = uniqueMIDINotes(in: currentStep)
-        let wrongMIDINotes = makeWrongCandidateMIDINotes(expectedMIDINotes)
-        audioRecognitionService.configureDetectorMode(
-            practiceAudioRecognitionDetectorModeSnapshot,
-            profile: harmonicTemplateTuningProfileSnapshot
-        )
-        audioStepAttemptAccumulator.setMode(.lowLatency)
-        audioRecognitionGeneration += 1
-        audioStepAttemptAccumulator.resetForNewStep(generation: audioRecognitionGeneration)
-
-        if isAudioRecognitionRunning {
-            audioRecognitionService.updateExpectedNotes(
-                expectedMIDINotes,
-                wrongCandidateMIDINotes: wrongMIDINotes,
-                generation: audioRecognitionGeneration
-            )
-            applyPendingAudioRecognitionSuppressIfNeeded(generation: audioRecognitionGeneration)
-            decisionLogger.debug("audio generation update=\(self.audioRecognitionGeneration, privacy: .public)")
-            return
-        }
-
-        isAudioRecognitionRunning = true
-        let startGeneration = audioRecognitionGeneration
-        let startExpectedMIDINotes = expectedMIDINotes
-        let startWrongMIDINotes = wrongMIDINotes
-        Task { @MainActor [weak self] in
-            guard let self else { return }
-            do {
-                try await audioRecognitionService.start(
-                    expectedMIDINotes: startExpectedMIDINotes,
-                    wrongCandidateMIDINotes: startWrongMIDINotes,
-                    generation: startGeneration,
-                    suppressUntil: audioRecognitionSuppressUntil.flatMap { $0 > Date() ? $0 : nil }
-                )
-                guard audioRecognitionGeneration == startGeneration,
-                      autoplayState == .off,
-                      isPracticeAudioRecognitionEnabled,
-                      case .guiding = state
-                else {
-                    stopAudioRecognition()
-                    return
-                }
-                applyPendingAudioRecognitionSuppressIfNeeded(generation: startGeneration)
-                decisionLogger.info("audio service started generation=\(startGeneration, privacy: .public)")
-            } catch {
-                isAudioRecognitionRunning = false
-                decisionLogger.error("audio service failed start generation=\(startGeneration, privacy: .public)")
-                recordAudioRecognitionError(error)
-            }
-        }
-    }
-
-    func refreshAudioRecognitionFromSettings() {
-        practiceAudioRecognitionDetectorModeSnapshot = Self.readPracticeAudioRecognitionDetectorMode()
-        harmonicTemplateTuningProfileSnapshot = Self.profile(for: practiceAudioRecognitionDetectorModeSnapshot)
-        audioRecognitionService?.configureDetectorMode(
-            practiceAudioRecognitionDetectorModeSnapshot,
-            profile: harmonicTemplateTuningProfileSnapshot
-        )
-        refreshAudioRecognitionForCurrentState()
-    }
-
-    func stopAudioRecognition() {
-        guard let audioRecognitionService else { return }
-        audioRecognitionGeneration += 1
-        audioStepAttemptAccumulator.resetForNewStep(generation: audioRecognitionGeneration)
-        audioRecognitionService.stop()
-        isAudioRecognitionRunning = false
-        audioRecognitionStatus = .stopped
-        decisionLogger.debug("audio service stopped by lifecycle")
-    }
-
-    private func prepareAudioRecognitionSuppressWindowForPlayback() -> Date {
-        let suppressUntil = Date().addingTimeInterval(audioRecognitionSuppressDuration)
-        audioRecognitionSuppressUntil = suppressUntil
-        audioRecognitionService?.suppressRecognition(
-            until: suppressUntil,
-            generation: audioRecognitionGeneration
-        )
-        return suppressUntil
-    }
-
-    private func applyPendingAudioRecognitionSuppressIfNeeded(generation: Int) {
-        guard let audioRecognitionService else { return }
-        guard let audioRecognitionSuppressUntil else { return }
-        guard audioRecognitionSuppressUntil > Date() else { return }
-        audioRecognitionService.suppressRecognition(
-            until: audioRecognitionSuppressUntil,
-            generation: generation
-        )
-    }
-
-    private func handleAudioRecognitionEvent(_ event: DetectedNoteEvent) {
-        guard isPracticeAudioRecognitionEnabled else { return }
-        guard autoplayState == .off else { return }
-        guard isManualReplayPlaying == false else { return }
-        guard event.generation == audioRecognitionGeneration else { return }
-        if let audioRecognitionSuppressUntil, event.timestamp <= audioRecognitionSuppressUntil {
-            decisionLogger.debug("audio event suppressed generation=\(event.generation, privacy: .public)")
-            return
-        }
-        guard let currentStep else { return }
-
-        let expectedMIDINotes = uniqueMIDINotes(in: currentStep)
-        let wrongMIDINotes = Set(makeWrongCandidateMIDINotes(expectedMIDINotes))
-
-        audioStepAttemptAccumulator.register(event: event)
-        let matchResult = audioStepAttemptAccumulator.evaluate(
-            expectedMIDINotes: expectedMIDINotes,
-            wrongCandidateMIDINotes: wrongMIDINotes,
-            generation: audioRecognitionGeneration,
-            at: event.timestamp,
-            handGateBoost: handGateState.isNearKeyboard || handGateState.hasDownwardMotion
-        )
-
-        switch matchResult {
-            case .matched:
-                audioStepAttemptAccumulator.markMatchedAndRequireRearm(
-                    expectedMIDINotes: expectedMIDINotes,
-                    at: event.timestamp
-                )
-                setFeedback(.correct)
-                advanceToNextStep()
-                decisionLogger.debug("audio matched advanced generation=\(event.generation, privacy: .public)")
-            case .wrong:
-                setFeedback(.wrong)
-                decisionLogger.debug("audio wrong feedback generation=\(event.generation, privacy: .public)")
-            case .insufficient:
-                break
-        }
-    }
-
-    private func makeWrongCandidateMIDINotes(_ expectedMIDINotes: [Int]) -> [Int] {
-        var result: Set<Int> = []
-        for note in expectedMIDINotes {
-            result.insert(note - 2)
-            result.insert(note - 1)
-            result.insert(note + 1)
-            result.insert(note + 2)
-        }
-        result.subtract(expectedMIDINotes)
-        return result.sorted()
-    }
-
-    private var isPracticeAudioRecognitionEnabled: Bool {
-        audioRecognitionEnabledSnapshot
-    }
-
-
-    var audioRecognitionGenerationForTesting: Int {
-        audioRecognitionGeneration
-    }
-
-    var audioRecognitionSuppressRemainingSeconds: TimeInterval {
-        guard let audioRecognitionSuppressUntil else { return 0 }
-        return max(0, audioRecognitionSuppressUntil.timeIntervalSinceNow)
-    }
-
-    private static func readPracticeAudioRecognitionDetectorMode() -> PracticeAudioRecognitionDetectorMode {
-        if let rawValue = UserDefaults.standard.string(forKey: "practiceStep3AudioRecognitionMode"),
-           let mode = PracticeAudioRecognitionDetectorMode(rawValue: rawValue)
-        {
-            return mode
-        }
-        return .harmonicTemplate
-    }
-
-    private static func profile(for _: PracticeAudioRecognitionDetectorMode) -> HarmonicTemplateTuningProfile {
-        .lowLatencyDefault
     }
 }
